Week8 JWT
didn't do as much custom work as I'd like to. I did make a custom folder in this lesson and began to structure it, but in the interest of time I want to continue to get the assignments done and submitted so I have time to really focus on the final project!